+++ xla/python/ifrt_proxy/common/array_util.cc
@@ -192,7 +192,7 @@ absl::StatusOr<std::vector<absl::Cord>> DeserializeStringHostBufferFromString(
 }
 
 absl::Status DeserializeFromCordIntoPreallocatedStringHostBuffer(
-    const absl::Cord& serialized_string_buffer, size_t num_elements,
+    const absl::Cord& serialized_string_buffer, int64_t num_elements,
     absl::Cord* preallocated_buffer) {
   proto::StringArrayContents string_array_proto;
 
@@ -211,7 +211,7 @@ absl::Status DeserializeFromCordIntoPreallocatedStringHostBuffer(
   // element so an inconsistent buffer cannot write past the end of
   // `preallocated_buffer`. This mirrors the element-count validation on the
   // server-side string deserialization path.
-  if (static_cast<size_t>(string_array_proto.strings_size()) != num_elements) {
+  if (string_array_proto.strings_size() != num_elements) {
     return absl::InvalidArgumentError(absl::StrCat(
         "String host buffer has ", string_array_proto.strings_size(),
         " elements but shape requires ", num_elements, " elements"));
